+++ tests/Unit/Media/Facade/MediaFacadeTest.php
@@ -10,19 +10,22 @@ declare(strict_types=1);
 namespace OxidEsales\MediaLibrary\Tests\Unit\Media\Facade;
 
 use OxidEsales\MediaLibrary\Media\DataType\MediaInterface;
+use OxidEsales\MediaLibrary\Media\DataType\MediaLookupContextInterface;
+use OxidEsales\MediaLibrary\Media\Exception\MediaNotFoundException;
 use OxidEsales\MediaLibrary\Media\Facade\MediaFacade;
 use OxidEsales\MediaLibrary\Media\Facade\MediaFacadeInterface;
 use OxidEsales\MediaLibrary\Media\Repository\PreloadMediaRepositoryInterface;
 use OxidEsales\MediaLibrary\Media\Service\MediaObjectResourceInterface;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
+use Psr\Log\LoggerInterface;
 
 class MediaFacadeTest extends TestCase
 {
     #[Test]
     public function registerForPreloadTriggersRepositoryMethod(): void
     {
-        $ids = [uniqid(), uniqid()];
+        $ids = [uniqid('mediaId'), uniqid('mediaId')];
 
         $preloadRepositorySpy = $this->createMock(PreloadMediaRepositoryInterface::class);
         $preloadRepositorySpy->expects($this->once())
@@ -39,15 +42,20 @@ class MediaFacadeTest extends TestCase
     #[Test]
     public function getsMediaFromRepository(): void
     {
-        $id = uniqid();
+        $id = uniqid('mediaId');
 
         $repositoryMock = $this->createMock(PreloadMediaRepositoryInterface::class);
         $repositoryMock->method('getMediaById')
             ->with($id)
             ->willReturn($mediaStub = $this->createStub(MediaInterface::class));
 
+        $loggerSpy = $this->createMock(LoggerInterface::class);
+        $loggerSpy->expects($this->never())
+            ->method('warning');
+
         $sut = $this->getSut(
             preloadMediaRepository: $repositoryMock,
+            logger: $loggerSpy,
         );
 
         $result = $sut->getMedia($id);
@@ -58,7 +66,7 @@ class MediaFacadeTest extends TestCase
     #[Test]
     public function calculatesTheUrlFromRepositoryMedia(): void
     {
-        $id = uniqid();
+        $id = uniqid('mediaId');
 
         $repositoryMock = $this->createMock(PreloadMediaRepositoryInterface::class);
         $repositoryMock->method('getMediaById')
@@ -68,11 +76,16 @@ class MediaFacadeTest extends TestCase
         $mediaObjectResourceMock = $this->createMock(MediaObjectResourceInterface::class);
         $mediaObjectResourceMock->method('getUrlToMedia')
             ->with($mediaStub)
-            ->willReturn($exampleUrl = uniqid());
+            ->willReturn($exampleUrl = uniqid('mediaUrl'));
+
+        $loggerSpy = $this->createMock(LoggerInterface::class);
+        $loggerSpy->expects($this->never())
+            ->method('warning');
 
         $sut = $this->getSut(
             preloadMediaRepository: $repositoryMock,
             mediaObjectResource: $mediaObjectResourceMock,
+            logger: $loggerSpy,
         );
 
         $result = $sut->getMediaUrl($id);
@@ -80,16 +93,121 @@ class MediaFacadeTest extends TestCase
         $this->assertSame($exampleUrl, $result);
     }
 
+    #[Test]
+    public function getMediaWithoutContextLogsAWarningAndRethrowsIfMediaNotFound(): void
+    {
+        $id = uniqid('mediaId');
+
+        $repositoryMock = $this->createMock(PreloadMediaRepositoryInterface::class);
+        $repositoryMock->method('getMediaById')
+            ->with($id)
+            ->willThrowException($exception = new MediaNotFoundException());
+
+        $loggerSpy = $this->createMock(LoggerInterface::class);
+        $loggerSpy->expects($this->once())
+            ->method('warning')
+            ->with(
+                'Media not found.',
+                [
+                    'mediaId' => $id,
+                    'trigger' => '',
+                    'identifier' => '',
+                    'note' => '',
+                ]
+            );
+
+        $sut = $this->getSut(
+            preloadMediaRepository: $repositoryMock,
+            logger: $loggerSpy,
+        );
+
+        $this->expectExceptionObject($exception);
+        $sut->getMedia($id);
+    }
+
+    #[Test]
+    public function getMediaUrlWithoutContextLogsAWarningAndRethrowsIfMediaNotFound(): void
+    {
+        $id = uniqid('mediaId');
+
+        $repositoryMock = $this->createMock(PreloadMediaRepositoryInterface::class);
+        $repositoryMock->method('getMediaById')
+            ->with($id)
+            ->willThrowException($exception = new MediaNotFoundException());
+
+        $loggerSpy = $this->createMock(LoggerInterface::class);
+        $loggerSpy->expects($this->once())
+            ->method('warning')
+            ->with(
+                'Media not found.',
+                [
+                    'mediaId' => $id,
+                    'trigger' => '',
+                    'identifier' => '',
+                    'note' => '',
+                ]
+            );
+
+        $sut = $this->getSut(
+            preloadMediaRepository: $repositoryMock,
+            logger: $loggerSpy,
+        );
+
+        $this->expectExceptionObject($exception);
+        $sut->getMediaUrl($id);
+    }
+
+    #[Test]
+    public function lookupContextIsLoggedWithTheMediaNotFoundWarning(): void
+    {
+        $id = uniqid('mediaId');
+
+        $repositoryMock = $this->createMock(PreloadMediaRepositoryInterface::class);
+        $repositoryMock->method('getMediaById')
+            ->with($id)
+            ->willThrowException(new MediaNotFoundException());
+
+        $contextStub = $this->createConfiguredStub(MediaLookupContextInterface::class, [
+            'getTrigger' => $trigger = uniqid('trigger'),
+            'getIdentifier' => $identifier = uniqid('identifier'),
+            'getNote' => $note = uniqid('note'),
+        ]);
+
+        $loggerSpy = $this->createMock(LoggerInterface::class);
+        $loggerSpy->expects($this->once())
+            ->method('warning')
+            ->with(
+                'Media not found.',
+                [
+                    'mediaId' => $id,
+                    'trigger' => $trigger,
+                    'identifier' => $identifier,
+                    'note' => $note,
+                ]
+            );
+
+        $sut = $this->getSut(
+            preloadMediaRepository: $repositoryMock,
+            logger: $loggerSpy,
+        );
+
+        $this->expectException(MediaNotFoundException::class);
+        $sut->getMediaUrl($id, $contextStub);
+    }
+
     private function getSut(
         ?PreloadMediaRepositoryInterface $preloadMediaRepository = null,
         ?MediaObjectResourceInterface $mediaObjectResource = null,
+        ?LoggerInterface $logger = null,
     ): MediaFacadeInterface {
         $preloadMediaRepository ??= $this->createStub(PreloadMediaRepositoryInterface::class);
         $mediaObjectResource ??= $this->createStub(MediaObjectResourceInterface::class);
+        $logger ??= $this->createStub(LoggerInterface::class);
 
         return new MediaFacade(
             preloadMediaRepository: $preloadMediaRepository,
             mediaObjectResource: $mediaObjectResource,
+            logger: $logger,
         );
     }
 }